requirements.txt
reno
pbr
flask
flask-cors
requests
PyYAML
